fix - make code cleaner requests and controller movie/quotes

+++ app/Http/Controllers/MovieController.php
@@ -2,14 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Movie\StoreMovieRequest;
-use App\Http\Requests\Movie\UpdateMovieRequest;
-use App\Http\Resources\Movie\MoviesResource;
-use App\Http\Resources\Movie\MovieEditResource;
-use App\Http\Resources\Movie\MovieResource;
 use App\Models\Movie;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Storage;
+use App\Http\Resources\Movie\MovieResource;
+use App\Http\Resources\Movie\MoviesResource;
+use App\Http\Requests\Movie\StoreMovieRequest;
+use App\Http\Requests\Movie\UpdateMovieRequest;
+use App\Http\Resources\Movie\MovieEditResource;
 
 class MovieController extends Controller
 {
@@ -20,6 +20,7 @@ class MovieController extends Controller
 
 	public function getMovie(Movie $movie): JsonResponse
 	{
+		$this->authorize('view', $movie);
 		return response()->json(new MovieResource($movie));
 	}
 
@@ -28,25 +29,12 @@ class MovieController extends Controller
 		$thumbnailPath = $request->thumbnail->store('movies', 'public');
 
 		$movie = Movie::create([
-			'user_id' => auth()->id(),
-			'name'    => [
-				'en' => $request->name_en,
-				'ka' => $request->name_ka,
-			],
-			'director' => [
-				'en' => $request->director_en,
-				'ka' => $request->director_ka,
-			],
-			'description' => [
-				'en' => $request->description_en,
-				'ka' => $request->description_ka,
-			],
-			'release_year' => $request->release_year,
+			...$request->validated(),
+			'user_id'      => auth()->id(),
 			'thumbnail'    => '/storage/' . $thumbnailPath,
 		]);
 
-		$genreIds = explode(',', $request->genre_ids);
-		$movie->genres()->attach($genreIds);
+		$movie->genres()->sync($request->genre_ids);
 
 		return response()->json(new MoviesResource($movie));
 	}
@@ -58,20 +46,9 @@ class MovieController extends Controller
 
 	public function updateMovie(UpdateMovieRequest $request, Movie $movie): JsonResponse
 	{
+		$this->authorize('update', $movie);
 		$movie->update([
-			'name'    => [
-				'en' => $request->name_en,
-				'ka' => $request->name_ka,
-			],
-			'director' => [
-				'en' => $request->director_en,
-				'ka' => $request->director_ka,
-			],
-			'description' => [
-				'en' => $request->description_en,
-				'ka' => $request->description_ka,
-			],
-			'release_year' => $request->release_year,
+			...$request->validated(),
 		]);
 		if ($request->hasFile('thumbnail')) {
 			Storage::disk('public')->delete($movie->thumbnail);
@@ -80,14 +57,15 @@ class MovieController extends Controller
 				'thumbnail' => '/storage/' . $thumbnailPath,
 			]);
 		}
-		$genreIds = explode(',', request()->genre_ids);
-		$movie->genres()->sync($genreIds);
+
+		$movie->genres()->sync($request->genre_ids);
 
 		return response()->json(new MoviesResource($movie));
 	}
 
 	public function deleteMovie(Movie $movie): JsonResponse
 	{
+		$this->authorize('delete', $movie);
 		$movie->delete();
 
 		return response()->json(['message' => 'Movie deleted successfully']);
